feat: Add AddFileModal to Resource Hub
