+++ assets/js/features/ResourceHub/AddNewFiles.tsx
@@ -1,13 +1,12 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 
-import { ResourceHub, useCreateResourceHubFolder } from "@/models/resourceHubs";
-
+import { ResourceHub } from "@/models/resourceHubs";
 import { IconFile, IconFolder, IconUpload } from "@tabler/icons-react";
 import { OptionsButton } from "@/components/Buttons";
-import Modal from "@/components/Modal";
-import Forms from "@/components/Forms";
 import { Paths } from "@/routes/paths";
+import { AddFolderModal } from "./AddFolderModal";
+import { AddFileModal, useAddFile } from "./AddFiles";
 
 interface Props {
   resourceHub: ResourceHub;
@@ -18,6 +17,7 @@ interface Props {
 export function AddFilesButtonAndForms({ resourceHub, refresh, folderId }: Props) {
   const navigate = useNavigate();
   const [showAddFolder, setShowAddFolder] = useState(false);
+  const fileProps = useAddFile();
 
   const toggleShowAddFolder = () => setShowAddFolder(!showAddFolder);
   const navigateToNewDocument = () => navigate(Paths.resourceHubNewDocumentPath(resourceHub.id!, folderId));
@@ -30,7 +30,7 @@ export function AddFilesButtonAndForms({ resourceHub, refresh, folderId }: Props
           options={[
             { icon: IconFile, label: "Write a new document", action: navigateToNewDocument, testId: "new-document" },
             { icon: IconFolder, label: "Create a new folder", action: toggleShowAddFolder, testId: "new-folder" },
-            { icon: IconUpload, label: "Upload files", action: () => {}, testId: "upload-files" },
+            { icon: IconUpload, label: "Upload files", action: fileProps.showAddFilePopUp, testId: "upload-files" },
           ]}
           testId="add-options"
         />
@@ -43,60 +43,8 @@ export function AddFilesButtonAndForms({ resourceHub, refresh, folderId }: Props
         refresh={refresh}
         folderId={folderId}
       />
+
+      <AddFileModal {...fileProps} />
     </>
-  );
-}
-
-interface FormProps {
-  showForm: boolean;
-  toggleForm: () => void;
-  refresh: () => void;
-  resourceHub: ResourceHub;
-  folderId?: string;
-}
-
-function AddFolderModal({ resourceHub, showForm, toggleForm, refresh, folderId }: FormProps) {
-  const [post] = useCreateResourceHubFolder();
-
-  const form = Forms.useForm({
-    fields: {
-      name: "",
-      description: null,
-    },
-    validate: (addError) => {
-      if (!form.values.name) {
-        addError("name", "Name is required");
-      }
-    },
-    cancel: toggleForm,
-    submit: async () => {
-      await post({
-        resourceHubId: resourceHub.id,
-        folderId: folderId,
-        name: form.values.name,
-        description: JSON.stringify(form.values.description),
-      });
-      refresh();
-      toggleForm();
-      form.actions.reset();
-    },
-  });
-
-  return (
-    <Modal title="New folder" isOpen={showForm} hideModal={toggleForm}>
-      <Forms.Form form={form}>
-        <Forms.FieldGroup>
-          <Forms.TextInput label="Name" field="name" />
-          <Forms.RichTextArea
-            label="Description"
-            field="description"
-            mentionSearchScope={{ type: "none" }}
-            placeholder="Description..."
-          />
-        </Forms.FieldGroup>
-
-        <Forms.Submit cancelText="Cancel" />
-      </Forms.Form>
-    </Modal>
   );
 }
